get customer accounts

+++ customer-api/src/main/java/com/pacvue/h10/customer/api/domain/customer/service/CustomerService.java
@@ -10,5 +10,5 @@ public interface CustomerService {
     /**
      * Returns a list of accounts available to the user
      */
-    List<AccountDto> getAccounts(Long userId);
+    List<AccountDto> getAccounts(Integer userId);
 }
